添加继电器控制功能
1. 继电器控制使能充电电路
2. 另一个继电器是错误警报

+++ User/CAN/hy_can.h
@@ -118,7 +118,7 @@ typedef enum HY_CANTASK_STATE{
 typedef struct HY_CANMSG{
 	uint32_t frame_id;
 	uint8_t databyte[8];
-	uint8_t resendcounts;
+	uint32_t resendcounts;
 }hy_canmsg;
 
 typedef struct CanComStrcut{
